scripts/requirements.txt
openpyxl>=3.0.0
pandas>=1.3.0,<2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
supabase>=0.7.0,<1.0.0
numpy>=1.21.0,<1.22.0
